python-django/requirements.txt
Django
mssql-django
pyodbc
djangorestframework

python_version >= "3.9"
